move function download from container in a file
